Extract footer
